Create separate SecondMenu component
